+++ bin/npm-unpack.js
@@ -1,11 +1,14 @@
 #!/usr/bin/env node
 
-var unpack = require('../lib/index.js').unpack
+var unpack = require("../lib/index.js").unpack;
 
-unpack({
-  directory: process.cwd()
-}, function (err) {
-  if (err) {
-    console.error(err)
+unpack(
+  {
+    directory: process.cwd(),
+  },
+  function (err) {
+    if (err) {
+      console.error(err);
+    }
   }
-})
+);