+++ test/npm.test.js
@@ -1,39 +1,39 @@
-var chai = require('chai')
-var expect = chai.expect
+var chai = require("chai");
+var expect = chai.expect;
 
-var rimraf = require('rimraf')
-var path = require('path')
-var npm = require('../lib/npm')
+var rimraf = require("rimraf");
+var path = require("path");
+var npm = require("../lib/npm");
 
-describe('npm', function () {
-  this.timeout(10000)
+describe("npm", function () {
+  this.timeout(10000);
 
-  describe('#pack', function () {
+  describe("#pack", function () {
     beforeEach(function () {
-      this.modulePath = path.resolve('./test/fixtures/fake-package')
-    })
+      this.modulePath = path.resolve("./test/fixtures/fake-package");
+    });
 
     afterEach(function (done) {
-      rimraf(path.resolve('./*.tgz'), done)
-    })
+      rimraf(path.resolve("./*.tgz"), done);
+    });
 
-    it('returns the tarred package name', function (done) {
+    it("returns the tarred package name", function (done) {
       npm.pack(this.modulePath, function (err, tarFile) {
-        expect(err).to.equal(null)
+        expect(err).to.equal(null);
 
-        expect(tarFile).to.eql('fake-package-1.2.3.tgz')
+        expect(tarFile).to.eql("fake-package-1.2.3.tgz");
 
-        done()
-      })
-    })
+        done();
+      });
+    });
 
-    it('errors if file cannot be found', function (done) {
-      npm.pack('/tmp/not-a-module', function (err) {
-        expect(err).to.not.equal(null)
-        expect(err.message).to.contain('Command failed')
+    it("errors if file cannot be found", function (done) {
+      npm.pack("/tmp/not-a-module", function (err) {
+        expect(err).to.not.equal(null);
+        expect(err.message).to.contain("Command failed");
 
-        done()
-      })
-    })
-  })
-})
+        done();
+      });
+    });
+  });
+});