+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # unreleased
+
 _Breaking Changes_
-* Drop suppport for node < v12
+
+- Drop suppport for node < v12
 
 # 0.1.1
 
-* Use execFile instead of exec to unpack module (issue #22)
+- Use execFile instead of exec to unpack module (issue #22)